octa_backend/requirements.txt
# OCTA Backend Dependencies

# Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Image Processing
Pillow>=10.0.0
numpy>=1.24.0
albumentations>=1.3.0  # 【Fix: Underfitting】强数据增强库

# Deep Learning Framework
torch>=2.0.0
torchvision>=0.15.0

# Archive formats
py7zr>=0.21.0
rarfile>=4.2

# Utilities
python-multipart>=0.0.6
requests>=2.31.0

# Fix: 平台优化 - 新增依赖
# Scheduled Tasks（定时任务）
APScheduler>=3.10.0
